E MANDI
The main objective of this project is build a website which will help civilian , retailer ,whole seller and even the farmer to get the best from his inputs. with the help of this a farmer will be able to know the best value for his vegetable and will not be fooled by the marketers. it will help in keeping the transparency between the whole seller and retailer and also the selection for civilian for his requirement become easy .so this will help in eradicating black marketing and inflation .